Smol dings

+++ SW3Projekt/SW3Projekt/Models/TimesheetEntry.cs
@@ -18,7 +18,7 @@ namespace SW3Projekt.Models
         public string ProjectID { get; set; }
         public DateTime  StartTime { get; set; }
         public DateTime EndTime { get; set; }
-        public double BreakTime { get; set; } = 0.5f;
+        public float BreakTime { get; set; } = 0.5f;
 
         public virtual List<VismaEntry> vismaEntries { get; set; }
